Fix to remedy the missing csdatatools library. Temporarily moved it into the liiatools package to allow for easy access until main external library can be updated.
